Moving the fuzz tests and fixing build issues.
